Data endpoints: foundation for meta programmatic CRUD per table
