--- models/__init__.py
@@ -1,11 +1,0 @@
-from models.database import db
-from models.question_group_templates import QuestionGroupTemplate
-from models.choices import Choice
-from models.forms import Form
-from models.questions import Question
-from models.choice_templates import ChoiceTemplate
-from models.madlibs import Madlib
-from models.question_templates import QuestionTemplate
-from models.clients import Client
-from models.question_groups import QuestionGroup
-from models.retail_chains import RetailChain
