+++ samples/test_db.py
@@ -1,6 +1,7 @@
 from sqlalchemy import create_engine
 from sqlalchemy.orm import scoped_session, sessionmaker
 
+from app.models import *
 
 client = Client(name = "L'Oreal")
 db.session.add(client)
